Remove check that the perf output file doesn't exist
Apparently, this file exists -- though who creates it hasn't been
thoroughly established, the suspicion is that we are running more than
once for most crates (which seems relatively likely). We'll go with the
old behavior though since presumably we always just used the last run.

If this turns out to not be the case we can investigate further.