Force deterministic rustc versioning
This changes our shim such that `rustc -vV` lies to Cargo, giving it canonical
output independent of the toolchain we're using.

Notably, this output feeds into the -Cmetadata hash and ultimately that affects
both incremental hashing (and as such what query order) and codegen unit
partitioning. This should make both more deterministic.

We may also need to rebuild std with this metadata hash.