Overlap upload with collection
Uploading is cheap CPU-time wise and shouldn't affect variability much. It'll
also only happen during self-profile collection anyway, which is pretty variable
as it is.

In the future we may experiment with alternative upload strategies, but for now
our primary goal is to eliminate the cost of uploads on run duration.